Внесены поправки в соответствии с рекомендациями по результатам ревью выполненного задания 7-го спринта.
