+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -12,7 +12,6 @@ import UIKit
 ///     - question: Текст вопроса
 ///     - questionNumber: Номер вопроса в квизе
 struct QuizStepViewModel {
-    
     let image: UIImage
     let question: String
     let questionNumber: String
