+++ MovieQuiz/Services/QuestionFactory.swift
@@ -7,6 +7,22 @@
 
 import Foundation
 
+// MARK: - QuestionFactoryProtocol
+
+protocol QuestionFactoryProtocol {
+    func requestNextQuestion()
+    func loadData()
+}
+
+// MARK: - QuestionFactoryDelegate
+
+protocol QuestionFactoryDelegate: AnyObject {
+    func didReceiveNextQuestion(question: QuizQuestion?)
+    func didLoadDataFromServer()
+    func didFailToLoadData(alert model: AlertModel)
+    func showLoadingIndicator(is: Bool)
+}
+
 // MARK: - Фабрика вопросов
 
 final class QuestionFactory: QuestionFactoryProtocol {
@@ -52,6 +68,7 @@ final class QuestionFactory: QuestionFactoryProtocol {
                 return
             }
             
+            // Задаём параметры нового вопроса
             let rating = Float(movie.rating) ?? 0
             let text = "Рейтинг этого фильма больше чем 7?"
             let correctAnswer = rating > 7
@@ -65,23 +82,24 @@ final class QuestionFactory: QuestionFactoryProtocol {
         }
     }
     
-    /// Loading movies data from server
+    ///  Загружаем данные о фильмах с сервера
     func loadData(){
         
         let moviesLoader = MoviesLoader()
         
         moviesLoader.loadMovies { [weak self] result in
+            
             DispatchQueue.main.async {
                 guard let self = self else { return }
                 
                 switch result {
+                    
                 case .success(let moviesHeap):
                     self.movies = moviesHeap.items
                     self.delegate?.didLoadDataFromServer()
-                case .failure(let error):
-//                    self.delegate?.didFailToLoadData(with: error)
-                    let messageText = error.localizedDescription.isEmpty ? "Не удалось загрузить данные с сервера" : error.localizedDescription
                     
+                case .failure(let error):
+                    let messageText = error.localizedDescription.isEmpty ? "Не удалось загрузить данные с сервера" : error.localizedDescription
                     let alertModel = AlertModel(title: "Ошибка", message: messageText, buttonText: "Попробовать ещё раз") { [weak self] _ in
                         self?.delegate?.showLoadingIndicator(is: true)
                         self?.loadData()
@@ -93,19 +111,3 @@ final class QuestionFactory: QuestionFactoryProtocol {
     }
 }
 
-// MARK: - QuestionFactoryProtocol
-
-protocol QuestionFactoryProtocol {
-    func requestNextQuestion()
-    func loadData()
-}
-
-// MARK: - QuestionFactoryDelegate
-
-protocol QuestionFactoryDelegate: AnyObject {
-    var viewController: MovieQuizViewControllerProtocol? { get }
-    func didReceiveNextQuestion(question: QuizQuestion?)
-    func didLoadDataFromServer()
-    func didFailToLoadData(alert model: AlertModel)
-    func showLoadingIndicator(is: Bool)
-}
